Footer: restore the usual look'n feel for Textual >= 0.63.0.
Textual 0.63.0 introduces a new Footer class that brings a significant visual
change. It also brings a ClassicFooter class for those who want to restore the
usual look'n feel but ClassicFooter is not meant to be perennial (it should be
dropped in 1.0.0).
This commit introduces a Footer override that restores the previous appearance.
This override is compatible with both old and new Footer classes.